feat: Created report_github_actions.py for GitHub Actions
Created the `report_github_actions.py` script to be used with GitHub Actions. This script integrates with GitHub Secrets, and all necessary secret files are created and managed using GitHub Secrets.

Updated the .gitignore file to ensure the exclusion of all data output, the initial script, and the credentials files from version control.